+++ apps/api/next.config.js
@@ -1,0 +1,96 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  transpilePackages: ["@floe/db", "@floe/utils"],
+
+  async headers() {
+    return [
+      {
+        /**
+         * CORS confif for client-side API requests (mostly needed for reactions currently)
+         */
+        source: "/v:version/reactions/:path*",
+        headers: [
+          { key: "Access-Control-Allow-Credentials", value: "true" },
+          //
+          { key: "Access-Control-Allow-Origin", value: "*" },
+          {
+            key: "Access-Control-Allow-Methods",
+            value: "GET,DELETE,PATCH,POST,PUT",
+          },
+          {
+            key: "Access-Control-Allow-Headers",
+            value:
+              "X-CSRF-Token, X-Requested-With, Accept, Accept-Version, Content-Length, Content-MD5, Content-Type, Date, X-Api-Version",
+          },
+        ],
+      },
+    ];
+  },
+
+  async rewrites() {
+    return {
+      /**
+       * Redirects. For more info see: https://stackoverflow.com/questions/65149521/managing-api-versions-using-nextjs-and-vercel
+       */
+      afterFiles: [
+        // This redirects requests recieved at / the root to the /api/ folder.
+        {
+          source: "/v:version/:rest*",
+          destination: "/api/v:version/:rest*",
+        },
+        // This redirects requests to api/v*/ to /api/ passing version as a query parameter.
+        {
+          source: "/api/v:version/:rest*",
+          destination: "/api/:rest*?version=:version",
+        },
+      ],
+      fallback: [
+        // These rewrites are checked after both pages/public files
+        // and dynamic routes are checked
+        {
+          source: "/:path*",
+          destination: `/api/:path*`,
+        },
+      ],
+    };
+  },
+};
+
+module.exports = nextConfig;
+
+// Injected content via Sentry wizard below
+
+const { withSentryConfig } = require("@sentry/nextjs");
+
+module.exports = withSentryConfig(
+  module.exports,
+  {
+    // For all available options, see:
+    // https://github.com/getsentry/sentry-webpack-plugin#options
+
+    // Suppresses source map uploading logs during build
+    silent: true,
+
+    org: "floe-641a5977f",
+    project: "javascript-nextjs",
+  },
+  {
+    // For all available options, see:
+    // https://docs.sentry.io/platforms/javascript/guides/nextjs/manual-setup/
+
+    // Upload a larger set of source maps for prettier stack traces (increases build time)
+    widenClientFileUpload: true,
+
+    // Transpiles SDK to be compatible with IE11 (increases bundle size)
+    transpileClientSDK: true,
+
+    // Routes browser requests to Sentry through a Next.js rewrite to circumvent ad-blockers (increases server load)
+    tunnelRoute: "/monitoring",
+
+    // Hides source maps from generated client bundles
+    hideSourceMaps: true,
+
+    // Automatically tree-shake Sentry logger statements to reduce bundle size
+    disableLogger: true,
+  }
+);