+++ packages/utils/src/files/fileNameToSlug.ts
@@ -1,0 +1,5 @@
+export const filenameToSlug = (filename: string) => {
+  return filename.replace(".floe/", "");
+};
+
+export const slugToFilename = (slug: string) => ".floe/" + slug;