+++ packages/utils/src/markdoc/index.ts
@@ -1,0 +1,2 @@
+export * from "./config";
+export { default as Markdoc } from "@markdoc/markdoc";